pruebas bitalino

+++ src/BITalino/BITalino.java
@@ -436,7 +436,7 @@ public class BITalino {
 		}
 	}
 	
-	public Frame[] read(double nSamples) throws BITalinoException 
+	public Frame[] read(int nSamples) throws BITalinoException 
 	{
                /** Reads acquisition frames from the device.
                 * This method returns when all requested frames are received from the device, or when a receive timeout occurs.
